feat(user); add delete confirmation to users

+++ frontend/src/app/constants/userTableColumns.tsx
@@ -56,8 +56,13 @@ const userTableColumns: ColumnDef<userType>[] = [
           <DropdownMenuContent align="end">
             <DropdownMenuItem
               onClick={() => {
-                UserService.delete("", user.id);
-                window.location.reload();
+                const confirmDelete = window.confirm(
+                  "Tem certeza que deseja deletar?"
+                );
+                if (confirmDelete) {
+                  UserService.delete("", user.id);
+                  window.location.href = "/machines";
+                }
               }}
               className="flex items-center justify-evenly gap-4"
             >
